requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
django-anymail==8.5
django-environ==0.8.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==6.0.0
